menu page

+++ backend/db/init.sql
@@ -7,7 +7,7 @@ DROP TABLE IF EXISTS food_order;
 CREATE TABLE IF NOT EXISTS food_menu (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price INTEGER NOT NULL,
     image_url VARCHAR(255),
     description TEXT
 );
